Task1/Пайплайн/requirements.txt
tensorflow>=2.10.0
keras>=2.10.0
numpy>=1.22.0
matplotlib>=3.5.0
pillow>=9.0.0
flask>=2.0.0
requests>=2.27.0
pytest>=7.0.0
scikit-learn>=1.0.0
